Correção e funcionamento do cadastro e consulta do cliente (parcial)

+++ GAME_WAVE - SISTEMA ASP_NET/Models/Classe_Cliente.cs
@@ -7,13 +7,13 @@ namespace GAME_WAVE___SISTEMA_ASP_NET.Models
 {
     public class Classe_Cliente
     {
-        public long Cli_cpf { get; set; }
+        public decimal Cli_cpf { get; set; }
 
         public string Cli_nome { get; set; }
 
         public string Cli_email { get; set; }
 
-        public long Cli_tel { get; set; }
+        public decimal Cli_tel { get; set; }
 
         public string Nome_uf { get; set; }
